remove old contouring method, intersecting method and rename new one
